Fix merge gc tests
